Support adding group in message from diff
Ading nodes in sequenced order so that the new node could be constructed
as in source node. Modified data types to hold temporary insertions and
deletions in accumulator used in editscript conversion. Hold the order
of appearance of insert node operation.

Support adding group, with preserving current group in accumulator.

Signed-off-by: 8c6794b6 <8c6794b6@gmail.com>