Update pattern expression classes in Scratch
Modified expression classes to not use type class constraints in result
type in object language.

Modified plam to take type of argument. This type is used in
deserialization, in fromTree function. From above change, all types are
fixed, so it should be possible to detect the type of argument used in
plam expression, but type detecting function is not yet written. In that
case, callee of the function will determine the type of lambda body, for
instance, id function does not know the actual resulst type until it is
called.

Signed-off-by: 8c6794b6 <8c6794b6@gmail.com>